requirements.txt
fastapi
uvicorn[standard]
streamlit
google-generativeai 
python-dotenv
pydantic
python-multipart
requests